+++ services/workflows-service/src/collection-flow/collection-flow.service.intg.test.ts
@@ -40,6 +40,9 @@ import { env } from '@/env';
 import { MerchantMonitoringClient } from '@/merchant-monitoring/merchant-monitoring.client';
 import { AnalyticsService } from '@/common/analytics-logger/analytics.service';
 import { WorkflowLogService } from '@/workflow/workflow-log.service';
+import { WorkflowRuntimeDataActorService } from '@/workflow/workflow-runtime-data-actor.service';
+import { ClsModule } from 'nestjs-cls';
+import { mockClsService } from '@/test/helpers/cls-service-helper';
 import { CollectionFlowStateService } from './collection-flow-state.service';
 import { DocumentService } from '@/document/document.service';
 
@@ -161,8 +164,11 @@ describe('CollectionFlowService', () => {
         EndUserRepository,
         MerchantMonitoringClient,
         WorkflowLogService,
+        WorkflowRuntimeDataActorService,
+        mockClsService(),
         CollectionFlowStateService,
       ],
+      imports: [ClsModule],
     }).compile();
 
     prismaClient = module.get<PrismaService>(PrismaService);